More Translations for MirahezeMagic Extension (#25)
* Translation Imported from Wikimedia

Portuguese Translation Imported from Wikimedia.

* Italian Translation

Italian Translation for MirahezeMagic Extension.

* Catalan Translation

Catalan Translation for MirahezeMagic Extension.

* Portuguese (Brasil) Translation

Portuguese (Brasil) Translation for MirahezeMagic Extension.

* Polish Translation

Polish Translation for MirahezeMagic Extension.

* Addition of GENDER in Permissions Translations

Addition of GENDER in Permissions Translations to distinguish the gender of the user who is a member of those groups.

* Russian Translation

Russian Translation for MirahezeMagic Extension.

* Addition of GENDER in Permissions Translations

Addition of GENDER in Permissions Translations to distinguish the gender of the user who is a member of those groups. This is a required requirement to aspire in the future to translations from translatewiki.net.

* Addition of GENDER in Permissions Translations

Addition of GENDER in Permissions Translations to distinguish the gender of the user who is a member of those groups. This is a required requirement to aspire in the future to translations from translatewiki.net.

* Addition of GENDER in Permissions Translations

Addition of GENDER in Permissions Translations to distinguish the gender of the user who is a member of those groups. This is a required requirement to aspire in the future to translations from translatewiki.net.

* Addition of GENDER in Permissions Translations

Addition of GENDER in Permissions Translations to distinguish the gender of the user who is a member of those groups. This is a required requirement to aspire in the future to translations from translatewiki.net.

* Addition of GENDER in Permissions Translations

Addition of GENDER in Permissions Translations to distinguish the gender of the user who is a member of those groups. This is a required requirement to aspire in the future to translations from translatewiki.net.

* Chinese Translation

Chinese Translation  for MirahezeMagic Extension.