Fix read/write acl on renamed containers
We call out to setContainersAccess which calls ->prepare. But within that function it checks if a container exists already and if it does, it does nothing. So no permissions are changed.

We fix this by manually setting each container that is renamed to use the correct acts.

Even so that maintenance script only did the core containers (no extension containers).