Add a warning regarding where to register the authenticator (#417)
WebAuthn authenticators only work on the specific domain where they were
registered. So, registering it on for example avid.miraheze.org, one may
find themselves surprised that they are not able to log-in at Meta if
they don't know how WebAuthn works.

Like in Wikimedia, we may want to warn users regarding this behavior.
Once logged in on the wiki where they registered the authenticator,
CentralAuth should take care of the rest and log them in on all the
various wikis as they browse.