Split HookHandlers by extension and some other cleanup
Also has major improvements to how wikis are filtered in the
GlobalUserPageWikis hook. Previously, it was making a connection to
ExtensionsModule in ManageWiki for every single wiki in databases.php.
That is now more than 20,000 wikis, which means, it was making 20,000
database connections every single time that hook ran. This has now been
changed to pull extensions directly from cache and check in there.